docs + move stuff + remove old stuff
