+++ pkg/model/modificator.go
@@ -1,8 +1,6 @@
 package model
 
 import (
-	"strings"
-
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -10,6 +8,7 @@ type ModificatorStatusReporter func(ModificatorStatus)
 
 type ModificatorController func(LocatorStatusStore) bool
 
+// ModificatorStatus is the status of the given Locator action after it's been attempted performed.
 type ModificatorStatus struct {
 	LocatorStatus
 	Error   error
@@ -19,8 +18,11 @@ type ModificatorStatus struct {
 }
 
 type ModificatorRegistrar func() (targetResourceType client.Object, modificator Modificator)
+
+// Modificator should perform the provided action on the given Resources provided by the Locator, e.g. Modify or Revert.
 type Modificator func(context SessionContext, ref Ref, store LocatorStatusStore, reporter ModificatorStatusReporter)
 
+// Sync is the entry point for ensuring the desired state for the given Ref is up to date.
 type Sync func(SessionContext, Ref, ModificatorController, LocatedReporter, ModificatorStatusReporter)
 
 // ModificatorStore Dummy impl for testing purposes.
@@ -30,18 +32,4 @@ type ModificatorStore struct {
 
 func (m *ModificatorStore) Report(status ModificatorStatus) {
 	m.Stored = append(m.Stored, status)
-}
-
-// ParseRefKindName parses a String() representation into a Object.
-func ParseRefKindName(exp string) RefKindName {
-	trimmedExp := strings.TrimSpace(strings.ToLower(exp))
-	parts := strings.Split(trimmedExp, "/")
-	if len(parts) == 2 {
-		return RefKindName{
-			Kind: parts[0],
-			Name: parts[1],
-		}
-	}
-
-	return RefKindName{Name: trimmedExp}
 }
